+++ CHANGES.md
@@ -12,6 +12,8 @@ Change Log
 * Add `cursor:pointer` to `Checkbox`
 * Use `yarn` in CI scripts (and upgrade node to v14)
 * `GeoJsonMixin.style["stroke-opacity"]` will now also set `polygonStroke.alpha`
+* Reduce `GeoJsonMixin` default stroke width from `2` to `1`
+* Add `TableMixin` styling to `GeoJsonMixin` - it will treat geojson feature properties as "rows" in a table - which can be styled in the same way as `TableMixin` (eg CSV). This is only enabled for geojson-vt/Protomaps (which requires `Terria.configParameters.enableGeojsonVt = true`). For more info see `GeojsonMixin.forceLoadMapItems()`
 * [The next improvement]
 
 #### 8.1.4